Remove letstune.Metric class
