+++ tests/results/test_epoch.py
@@ -7,7 +7,6 @@ import pandas as pd
 import pytest
 from pandas import Timedelta, Timestamp
 
-import letstune
 from letstune.backend import repo
 from letstune.backend.scheduler.epoch import Config
 from letstune.results import epoch
@@ -15,7 +14,7 @@ from letstune.results import epoch
 from . import utils
 from .utils import ModelParams, OptimizerParams
 
-METRIC = letstune.Metric("accuracy")
+METRIC = "accuracy"
 
 
 @pytest.fixture
